+++ src/IncrementalInference.jl
@@ -107,6 +107,9 @@ include("ExportAPI.jl")
 ## =============================
 # Source code
 
+# FIXME, move up to DFG
+# abstract type AbstractManifoldMinimize <: AbstractRelative end
+
 # regular
 include("entities/SolverParams.jl")